Flatten `core` module
Resolves #6 , #8
Finalizes rewrite milestone :)
