--- src/solvers/psr/core/filters/mod.rs
@@ -1,1 +1,0 @@
-pub mod known;
