--- src/solvers/psr/core/finders/mod.rs
@@ -1,5 +1,0 @@
-/// Single candidate detection.
-pub mod single_candidate;
-
-/// Unique candidate detection.
-pub mod unique_candidate;
