+++ src/solvers/psr/core/section.rs
@@ -1,17 +1,26 @@
+use super::{
+    cell::{CandidateError, Cell},
+    idx::{GridIdx, InnerIdx, SectionIdx},
+    num::Num,
+};
+
 mod simd;
 #[doc(inline)]
 pub use simd::*;
 
-use super::{
-    cell::Cell,
-    idx::{GridIdx, InnerIdx, SectionIdx},
-};
+// use structs::{
+//     cell::{CandidateError, Cell},
+//     grid::Grid,
+//     idx::{GridIdx, InnerIdx},
+//     num::Num,
+//     section::{Section, SectionInfo, SectionKind::*},
+// };
 
 #[derive(Debug, Clone)]
 /// A row, column, or box of [`Cells`](Cell).
 pub struct Section {
     /// The information linking to what [`Section`] within the
-    /// [`Grid`](super::super::structs::grid::Grid) this section was taken from.
+    /// [`Grid`](super::grid::Grid) this section was taken from.
     pub info: SectionInfo,
     /// The underlying array of [`Cells`](Cell).
     pub cells: [Cell; 9],
@@ -41,6 +50,44 @@ impl Section {
     /// Create a new [`Section`].
     pub(super) fn new(info: SectionInfo, cells: [Cell; 9]) -> Self {
         Self { info, cells }
+    }
+
+    // ========================================================
+    // Solving
+    // ========================================================
+
+    /// Finds if the `Cell` at `inner_idx` contains a candidate found in no
+    /// other cells within the `Section`.
+    ///
+    /// # Finder Method
+    /// See the [core module's documentation](super::super#finders) for more information.
+    ///
+    /// # Safety
+    /// Does not check that the inner `Cell`s have correct representation.
+    /// Incorrect `Cell` representations will lead to undefined behaviour.
+    pub fn unique_candidate(&self, inner_idx: InnerIdx) -> Result<Option<Num>, CandidateError> {
+        let mut cells = self.cells;
+        let candidate_cell = cells[inner_idx];
+
+        // OR all of the remaining candidates together
+        cells[inner_idx] = unsafe { Cell::zeroed() };
+        let combined_candidates = cells.iter().fold(unsafe { Cell::zeroed() }, |a, b| unsafe {
+            Cell::new_unchecked(a.to_u16() | b.to_u16())
+        });
+
+        let unique_candidates_u16 =
+            // Find bits set in candidate_cell but not in combined_candidates.
+            (candidate_cell.to_u16() & !combined_candidates.to_u16());
+
+        match unique_candidates_u16.count_ones() {
+            0 => Ok(None),
+            1 => {
+                let n = unique_candidates_u16.ilog2() as u8;
+                let num = unsafe { Num::new_unchecked(n) };
+                Ok(Some(num))
+            }
+            _ => Err(CandidateError::MultipleUniqueCandidates),
+        }
     }
 }
